:truck: tech(api): move candidate registration to session-management
