+++ api/src/certification/session-management/application/session-controller.js
@@ -61,11 +61,9 @@ async function getJuryCertificationSummaries(
   });
 }
 
-const sessionController = {
+export const sessionController = {
   getJuryCertificationSummaries,
   findPaginatedFilteredJurySessions,
   get,
   getJurySession,
 };
-
-export { sessionController };
