+++ api/src/certification/session-management/domain/services/reconcile-candidate.js
