+++ api/tests/certification/session-management/unit/application/certification-candidate-controller_test.js
@@ -3,6 +3,7 @@ import sinon from 'sinon';
 
 import { certificationCandidateController } from '../../../../../src/certification/session-management/application/certification-candidate-controller.js';
 import { usecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
+import { normalize } from '../../../../../src/shared/infrastructure/utils/string-utils.js';
 import { hFake } from '../../../../tooling/mocks/hapi.mock.js';
 
 describe('Certification | Session Management | Unit | Application | Controller | Certification Candidate', function () {
@@ -118,6 +119,69 @@ describe('Certification | Session Management | Unit | Application | Controller |
 
       // then
       sinon.assert.calledWithExactly(usecases.endAssessmentByInvigilator, { certificationCandidateId: 2 });
+    });
+  });
+
+  describe('#createCandidateParticipation', function () {
+    it('should return candidate information', async function () {
+      // given
+      const sessionId = 123;
+      const userId = 274939274;
+      const firstName = 'Jeanne';
+      const lastName = 'Serge';
+      const birthdate = '2020-10-10';
+
+      const request = {
+        payload: {
+          data: {
+            attributes: {
+              'first-name': firstName,
+              'last-name': lastName,
+              birthdate,
+            },
+          },
+        },
+        auth: { credentials: { userId } },
+        params: { sessionId },
+        headers: { origin: 'https://app.pix.fr' },
+      };
+      const candidate = {
+        firstName,
+        lastName,
+        birthdate,
+        sessionId,
+        hasSeenCertificationInstructions: false,
+      };
+
+      sinon.stub(usecases, 'registerCandidateParticipation').resolves();
+      usecases.registerCandidateParticipation
+        .withArgs({
+          userId,
+          sessionId,
+          firstName,
+          lastName,
+          birthdate,
+          isFrenchDomainExtension: true,
+          normalizeStringFnc: normalize,
+        })
+        .resolves(candidate);
+
+      // when
+      const response = await certificationCandidateController.createCandidateParticipation(request, hFake);
+
+      // then
+      expect(response.source).to.deep.equal({
+        data: {
+          attributes: {
+            birthdate,
+            'first-name': firstName,
+            'has-seen-certification-instructions': false,
+            'last-name': lastName,
+            'session-id': sessionId,
+          },
+          type: 'certification-candidates',
+        },
+      });
     });
   });
 });
