+++ api/tests/certification/session-management/unit/domain/usecases/register-candidate-participation_test.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { WrongDomainExtensionForPixPlusError } from '../../../../../../src/certification/enrolment/domain/errors.js';
-import { registerCandidateParticipation } from '../../../../../../src/certification/enrolment/domain/usecases/register-candidate-participation.js';
+import { WrongDomainExtensionForPixPlusError } from '../../../../../../src/certification/session-management/domain/errors.js';
+import { registerCandidateParticipation } from '../../../../../../src/certification/session-management/domain/usecases/register-candidate-participation.js';
 import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { UserNotAuthorizedToCertifyError } from '../../../../../../src/shared/domain/errors.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
